// nodejs/src/types.ts
export interface EphemeralJoinResponse {
  ephemeralId: string;
  roomId: string;
  token: string;
  inviteCode: string;
}

export interface EphemeralChatResponse {
  messages?: EphemeralChatMessage[];
  since?: string;
}

export interface EphemeralChatMessage {
  id: string;
  ephemeralId: string;
  content: string;
  speak: boolean;
  timestamp: string;
}